refactoring nom des dossiers camalCase
