+++ app/_layout.js
@@ -4,9 +4,9 @@ export default function RootLayout() {
   return (
     <Stack>
       <Stack.Screen name="index" options={{headerShown: false}} />
-      <Stack.Screen name="Page/JoueurPage" options={{headerShown: false}} />
-      <Stack.Screen name="Page/EditPage" options={{headerShown: false}} />
-      <Stack.Screen name="Page/GamePage" options={{headerShown: false, gestureEnabled: true }} />
+      <Stack.Screen name="pages/JoueurPage" options={{headerShown: false}} />
+      <Stack.Screen name="pages/EditPage" options={{headerShown: false}} />
+      <Stack.Screen name="pages/GamePage" options={{headerShown: false, gestureEnabled: true }} />
     </Stack>
   );
 }
